code/server/requirements.txt
flask
flask-mysql
pymysql
mysql-connector-python
cryptography